refactor: Rename Link-Converters to Link-Resolvers
